update pgs test

+++ sb_rdbms/postgres/src/main/java/com/example/pgs/demo/persistence/CustomerRepository.java
@@ -3,4 +3,4 @@ package com.example.pgs.demo.persistence;
 import com.example.pgs.demo.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> { }
+public interface CustomerRepository extends JpaRepository<Customer, Integer> { }
